+++ src/openorganisatie/organisatie/api/tests/test_organisatorische_eenheid.py
@@ -1,6 +1,5 @@
 from datetime import date, datetime
 
-from django.core.exceptions import ValidationError
 from django.urls import reverse
 from django.utils.timezone import make_aware
 
@@ -309,36 +308,6 @@ class OrganisatorischeEenheidAPITests(APITestCase):
         )
 
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_prevent_cycle_in_parent(self):
-        parent = OrganisatorischeEenheidFactory()
-        child = OrganisatorischeEenheidFactory(hoofd_organisatorische_eenheid=parent)
-
-        parent.hoofd_organisatorische_eenheid = child
-        with self.assertRaises(ValidationError) as val:
-            parent.clean()
-        self.assertIn(
-            "hoofd_organisatorische_eenheid",
-            val.exception.message_dict,
-        )
-        self.assertIn(
-            "Een organisatorische eenheid kan geen kind als bovenliggende eenheid hebben.",
-            val.exception.message_dict["hoofd_organisatorische_eenheid"][0],
-        )
-
-    def test_prevent_self_parenting(self):
-        org = OrganisatorischeEenheidFactory()
-        org.hoofd_organisatorische_eenheid = org
-        with self.assertRaises(ValidationError) as val:
-            org.clean()
-        self.assertIn(
-            "hoofd_organisatorische_eenheid",
-            val.exception.message_dict,
-        )
-        self.assertIn(
-            "Een organisatorische eenheid kan niet naar zichzelf verwijzen.",
-            val.exception.message_dict["hoofd_organisatorische_eenheid"][0],
-        )
 
     def test_history(self):
         url = reverse("organisatie_api:organisatorischeeenheid-list")
